refactor: extract pcb grid svg creation
